Move OneShot library search to top tab
