+++ tests/oneshot-workflows.test.tsx
@@ -8,7 +8,7 @@ import {
   setSavedBlueprintCollection,
 } from '../src/state/blueprints';
 import { createInspirationBoard, createInspirationPin } from '../src/state/inspiration';
-import type { DesignSystemSummary, Project, SkillSummary } from '../src/types';
+import type { DesignSystemSummary, SkillSummary } from '../src/types';
 
 function skill(id: string, mode: SkillSummary['mode']): SkillSummary {
   return {
@@ -39,23 +39,6 @@ function designSystem(id: string): DesignSystemSummary {
     category: 'Product',
     summary: `${id} design system`,
     swatches: [],
-  };
-}
-
-function project(id: string, name: string): Project {
-  return {
-    id,
-    name,
-    skillId: null,
-    designSystemId: null,
-    createdAt: 1000,
-    updatedAt: 3000,
-    metadata: {
-      kind: 'prototype',
-      workflowTitle: 'Dashboard Mockup',
-      workflowCategory: 'Product prototype',
-      workflowOutcome: 'Operational UI concept',
-    },
   };
 }
 
@@ -353,85 +336,6 @@ describe('OneShotWorkflows', () => {
           workflowId: 'oneshot-cover-run',
           workflowTitle: 'OneShot Cover Run',
         }),
-      }),
-    );
-  });
-
-  it('searches blueprints, boards, and projects from the unified library', () => {
-    const onCreateProject = vi.fn();
-    const onOpenProject = vi.fn();
-    saveWorkflowBlueprint({
-      metadata: {
-        kind: 'template',
-        workflowId: 'oneshot-cover-run',
-        workflowTitle: 'OneShot Cover Run',
-        workflowCategory: 'Book cover production',
-      },
-      prompt: 'Use the OneShot workflow blueprint: OneShot Cover Run.',
-      skillId: 'digital-eguide',
-      designSystemId: 'warm-editorial',
-    });
-    const board = createInspirationBoard({
-      title: 'Cover moodboard',
-      description: 'Publishing references for a cover.',
-      tags: ['cover'],
-    });
-    createInspirationPin({
-      boardId: board.id,
-      title: 'Cover source',
-      sourceUrl: 'local/cover-source.html',
-      note: 'Use the title hierarchy.',
-      tags: ['typography'],
-    });
-
-    render(
-      <OneShotWorkflows
-        skills={[skill('digital-eguide', 'template')]}
-        designSystems={[designSystem('warm-editorial')]}
-        defaultDesignSystemId="default"
-        projects={[project('project-1', 'Project archive')]}
-        onCreateProject={onCreateProject}
-        onOpenProject={onOpenProject}
-      />,
-    );
-
-    expect(screen.getByLabelText('OneShot library search')).toBeInTheDocument();
-    expect(screen.getByText('1 blueprints')).toBeInTheDocument();
-    expect(screen.getByText('3 boards')).toBeInTheDocument();
-    expect(screen.getByText('1 projects')).toBeInTheDocument();
-
-    fireEvent.change(screen.getByPlaceholderText('Search blueprints, boards, and projects'), {
-      target: { value: 'archive' },
-    });
-    fireEvent.click(screen.getByRole('button', { name: 'Open Project archive' }));
-    expect(onOpenProject).toHaveBeenCalledWith('project-1');
-
-    fireEvent.change(screen.getByPlaceholderText('Search blueprints, boards, and projects'), {
-      target: { value: 'moodboard' },
-    });
-    fireEvent.click(screen.getByRole('button', { name: 'Create brief from Cover moodboard' }));
-    expect(onCreateProject).toHaveBeenCalledWith(
-      expect.objectContaining({
-        name: 'Cover moodboard reference brief',
-        metadata: expect.objectContaining({
-          workflowReferenceBoardId: board.id,
-          workflowReferenceBoardTitle: 'Cover moodboard',
-          workflowReferencePinCount: 1,
-        }),
-        pendingPrompt: expect.stringContaining('Source: local/cover-source.html'),
-      }),
-    );
-
-    onCreateProject.mockClear();
-    fireEvent.change(screen.getByPlaceholderText('Search blueprints, boards, and projects'), {
-      target: { value: 'blueprint' },
-    });
-    fireEvent.click(screen.getByRole('button', { name: 'Start OneShot Cover Run' }));
-    expect(onCreateProject).toHaveBeenCalledWith(
-      expect.objectContaining({
-        name: 'OneShot Cover Run',
-        skillId: 'digital-eguide',
-        designSystemId: 'warm-editorial',
       }),
     );
   });
